# sql: introduce role-level audit logging