--- a/pkg/server/BUILD.bazel
+++ b/pkg/server/BUILD.bazel
@@ -196,6 +196,7 @@ go_library(
         "//pkg/spanconfig/spanconfigstore",
         "//pkg/sql",
         "//pkg/sql/appstatspb",
+        "//pkg/sql/auditlogging",
         "//pkg/sql/catalog",
         "//pkg/sql/catalog/bootstrap",
         "//pkg/sql/catalog/catalogkeys",