--- a/pkg/server/structlogging/BUILD.bazel
+++ b/pkg/server/structlogging/BUILD.bazel
@@ -45,6 +45,7 @@ go_test(
         "//pkg/util/leaktest",
         "//pkg/util/log",
         "//pkg/util/log/logcrash",
+        "//pkg/util/log/logpb",
         "//pkg/util/log/logtestutils",
         "//pkg/util/randutil",
         "@com_github_cockroachdb_errors//:errors",