--- a/pkg/settings/rulebasedscanner/BUILD.bazel
+++ b/pkg/settings/rulebasedscanner/BUILD.bazel
@@ -1,0 +1,28 @@
+load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "rulebasedscanner",
+    srcs = [
+        "scanned_input.go",
+        "scanner.go",
+    ],
+    importpath = "github.com/cockroachdb/cockroach/pkg/settings/rulebasedscanner",
+    visibility = ["//visibility:public"],
+    deps = ["@com_github_cockroachdb_errors//:errors"],
+)
+
+go_test(
+    name = "rulebasedscanner_test",
+    srcs = ["scanner_test.go"],
+    args = ["-test.timeout=295s"],
+    data = glob(["testdata/**"]),
+    embed = [":rulebasedscanner"],
+    deps = [
+        "//pkg/testutils/datapathutils",
+        "@com_github_cockroachdb_datadriven//:datadriven",
+        "@com_github_kr_pretty//:pretty",
+    ],
+)
+
+get_x_data(name = "get_x_data")