--- a/pkg/sql/auditlogging/BUILD.bazel
+++ b/pkg/sql/auditlogging/BUILD.bazel
@@ -2,32 +2,38 @@ load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "hba",
+    name = "auditlogging",
     srcs = [
-        "hba.go",
+        "audit_log.go",
+        "audit_log_config.go",
         "parser.go",
     ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/sql/pgwire/hba",
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/auditlogging",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/kv",
         "//pkg/security/username",
+        "//pkg/settings",
+        "//pkg/settings/cluster",
         "//pkg/settings/rulebasedscanner",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/sem/tree",
-        "//pkg/util/errorutil/unimplemented",
+        "//pkg/util/log",
+        "//pkg/util/log/eventpb",
+        "//pkg/util/log/logpb",
+        "//pkg/util/syncutil",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_olekukonko_tablewriter//:tablewriter",
     ],
 )
 
 go_test(
-    name = "hba_test",
-    size = "small",
-    srcs = ["hba_test.go"],
-    args = ["-test.timeout=55s"],
+    name = "auditlogging_test",
+    srcs = ["audit_log_test.go"],
+    args = ["-test.timeout=295s"],
     data = glob(["testdata/**"]),
-    embed = [":hba"],
+    embed = [":auditlogging"],
     deps = [
         "//pkg/settings/rulebasedscanner",
         "//pkg/testutils/datapathutils",