--- a/pkg/sql/auditlogging/auditevents/BUILD.bazel
+++ b/pkg/sql/auditlogging/auditevents/BUILD.bazel
@@ -1,0 +1,23 @@
+load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "auditevents",
+    srcs = [
+        "role_based_audit_event.go",
+        "sensitive_table_access.go",
+    ],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/auditlogging/auditevents",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/sql/auditlogging",
+        "//pkg/sql/catalog",
+        "//pkg/sql/sem/tree",
+        "//pkg/util/log",
+        "//pkg/util/log/eventpb",
+        "//pkg/util/log/logpb",
+        "@com_github_cockroachdb_errors//:errors",
+    ],
+)
+
+get_x_data(name = "get_x_data")