--- a/pkg/sql/conn_executor.go
+++ b/pkg/sql/conn_executor.go
@@ -3557,7 +3557,7 @@ func (ex *connExecutor) implicitTxn() bool {
 	return ok && os.ImplicitTxn.Get()
 }
 
-// initPlanner initializes a planner so it can can be used for planning a
+// initPlanner initializes a planner so it can be used for planning a
 // query in the context of this session.
 func (ex *connExecutor) initPlanner(ctx context.Context, p *planner) {
 	p.cancelChecker.Reset(ctx)