… docs/generated/eventlog.md
… pkg/BUILD.bazel
… pkg/ccl/telemetryccl/telemetry_logging_test.go
… pkg/server/structlogging/hot_ranges_log_test.go
… pkg/settings/rulebasedscanner/scanner_test.go
… pkg/sql/BUILD.bazel
… pkg/sql/audit_logging_test.go
… pkg/sql/auditlogging/audit_log_test.go
… pkg/sql/auditlogging/auditevents/sensitive_table_access.go
… pkg/sql/auditlogging/parser.go
… pkg/sql/auditlogging/testdata/parse
… pkg/sql/copy_test.go
… pkg/sql/exec_log.go
… pkg/sql/pgwire/hba/hba.go
… pkg/sql/pgwire/hba/hba_test.go
… pkg/sql/pgwire/hba/parser.go
… pkg/sql/pgwire/hba/testdata/parse
… pkg/sql/pgwire/hba_conf.go
… pkg/sql/pgwire/server.go
… pkg/sql/plan.go
… pkg/sql/planner.go
… pkg/sql/scheduledlogging/captured_index_usage_stats_test.go
… pkg/sql/sessioninit/BUILD.bazel
… pkg/sql/sessioninit/cache.go
… pkg/sql/telemetry_logging_test.go
… pkg/util/log/eventpb/eventlog_channels_generated.go
… pkg/util/log/eventpb/json_encode_generated.go
… pkg/util/log/eventpb/sql_audit_events.proto
… pkg/util/log/logtestutils/BUILD.bazel
… pkg/util/log/logtestutils/log_test_utils.go