Resolves: #100467

**Configuration**

This change introduces role-level audit logging through a cluster setting `sql.log.user_audit`.

Using the cluster setting, a user is able to set an audit logging configuration using a table-like syntax, like so:

```
set cluster setting sql.log.user_audit = '
    test_role       ALL
    another_role    NONE
    ...
    ALL             NONE
';
```

Each row in the configuration represents an _audit setting_ in the configuration. An audit setting is comprised of the following columns:
- `USER/ROLE`: the name of the user/role we want to audit (i.e. `test_role` in the example above)
    - the _special_ user/role value `ALL` applies the audit setting to *all* users

- `STATEMENT_FILTER`: the statement filter for the role (i.e. `ALL` or `NONE` in the example above)
    - `ALL` will enable audit logging for all SQL statements for the given audit setting
    - `NONE` will exclude audit logging for all SQL statements for the given audit setting

As such, the example above can be read as:
- users with the username/role `test_role` will emit audit logs for all statements they issue
- users with the username/role `another_role` will not emit audit logs for any statements they issue (in principle, one can achieve this by simply omitting `another_role` from the configuration entirely, but we include it here as an example)
- all remaining users will not emit audit logs for any statements they issue

Internally, this configuration is stored as an object containing a list of audit settings. The configuration is stored as part of the executor config, so it can be shared amongst all SQL connections.

**Matching**

To determine whether a statement should be emitted to the audit log, we check whether there is a corresponding audit setting for the statement. This means:
- the user who issued the statement must have a username or belong to a role that has a corresponding audit setting
- for that audit setting, check if auditing is enabled (i.e. the statement filter must be `ALL`)

To do this, we get the user's role memberships from the role membership cache and iterate through the config's audit settings, looking for a matching audit setting by role. *At the first audit setting match*, we stop iteration. We check if the auditing is enabled (i.e. statement filter must be `ALL`), if so, we add the statement to a collection of audit events to be logged later. If the statement filter is `NONE`, we stop there.

It's worth re-iterating that the *order in which the audit settings in the configuration are specified is important*, as the *first matching audit setting* will apply first.

Additionally, we make sure to filter out any statements issued by:
- reserved users
- internal planners
- internal executors
from the audit logs to avoid noise in the logs.

Release note: Introduce role-based auditing through new cluster setting `sql.log.user_audit`.

Using the cluster setting, a user is able to set an audit logging configuration using a table-like syntax:

```
set cluster setting sql.log.user_audit = '
    test_role       ALL
    another_role    NONE
    ALL             NONE
';
```

Each row in the configuration represents an _audit setting_ in the configuration. An audit setting is comprised of the following columns:
- `USER/ROLE`: the name of the user/role we want to audit (i.e. `test_role` in the example above)
    - the _special_ user/role value `ALL` applies the audit setting to *all* users

- `STATEMENT_FILTER`: the statement filter for the role (i.e. `ALL` or `NONE` in the example above)
    - `ALL` will enable audit logging for all SQL statements for the given audit setting
    - `NONE` will exclude audit logging for all SQL statements for the given audit setting

As such, the example above can be read as:
- users with the username/role `test_role` will emit audit logs for all statements they issue
- users with the username/role `another_role` will not emit audit logs for any statements they issue (in principle, one can achieve this by simply omitting `another_role` from the configuration entirely, but we include it here as an example)
- all remaining users will not emit audit logs for any statements they issue

**Matching**

To determine whether a statement should be emitted to the audit log, we check whether there is a corresponding audit setting for the statement. This means:
- the user who issued the statement must have a username or belong to a role that has a corresponding audit setting
- for that audit setting, check if auditing is enabled (i.e. the statement filter must be `ALL`)

*At the first audit setting match (based on role/username)*, we stop iteration. We check if the auditing is enabled for that audit setting (i.e. statement filter must be `ALL`), if so, we add the statement to a collection of audit events to be logged later. If the statement filter is `NONE`, we stop there. Consequently, the *order in which the audit settings in the configuration are specified is important*, as only the *first matching audit setting* will apply.

Note that this is an initial implementation (i.e. v1) and is subject to change/evolve going forward.