@THardy98 2023-04-28
Here are some benchmarks running TPCC for 20 minutes, 2500 warehouses on a 3 node cluster on `master` and with these changes. I tested these changes with 3 different audit configuration sizes (empty - no entries, small - 50 entries, large - 200 entries). The size of the config determines the number of audit settings we need to iterate over before we reach the matching audit setting for the user (more entries = more iterations).

The roachtest in this PR was used to run the TPCC workloads on these changes. Note that the TPCC run on `master` was also used with the `--wait=false --workers=2500` workload flags.

```
*** MASTER ***
**TWENTY MINUTES, 2500 WAREHOUSES**
_elapsed_______tpmC____efc__avg(ms)__p50(ms)__p90(ms)__p95(ms)__p99(ms)_pMax(ms)
 1200.0s    28699.7  89.3%   2619.1   2550.1   3892.3   4563.4   6442.5  23622.3
```

``` 
*** THIS CHANGE ***
**TWENTY MINUTES, 2500 WAREHOUSES**

Run 1
**EMPTY CONFIG**
_elapsed_______tpmC____efc__avg(ms)__p50(ms)__p90(ms)__p95(ms)__p99(ms)_pMax(ms)
 1200.0s    29130.4  90.6%   2572.2   2550.1   3892.3   4831.8   6979.3  20401.1
 
**SMALL (50 ENTRIES)**
_elapsed_______tpmC____efc__avg(ms)__p50(ms)__p90(ms)__p95(ms)__p99(ms)_pMax(ms)
 1200.0s    28210.2  87.7%   2608.2   2550.1   4026.5   4831.8   7247.8  21474.8
 
**LARGE (200 ENTRIES)**
_elapsed_______tpmC____efc__avg(ms)__p50(ms)__p90(ms)__p95(ms)__p99(ms)_pMax(ms)
 1200.0s    27516.7  85.6%   2662.3   2550.1   3892.3   4563.4   6174.0  19327.4

Run 2
**EMPTY CONFIG**
_elapsed_______tpmC____efc__avg(ms)__p50(ms)__p90(ms)__p95(ms)__p99(ms)_pMax(ms)
 1200.0s    29090.7  90.5%   2456.7   2415.9   3623.9   4295.0   6442.5  20401.1

**SMALL (50 ENTRIES)**
_elapsed_______tpmC____efc__avg(ms)__p50(ms)__p90(ms)__p95(ms)__p99(ms)_pMax(ms)
 1200.0s    28843.3  89.7%   2528.6   2550.1   3623.9   4026.5   5100.3  19327.4

**LARGE (200 ENTRIES)**
 _elapsed_______tpmC____efc__avg(ms)__p50(ms)__p90(ms)__p95(ms)__p99(ms)_pMax(ms)
 1200.0s    28560.1  88.8%   2609.3   2550.1   3623.9   4160.7   5905.6  25769.8
```

The variation between the 2 runs with these changes makes it difficult to actually determine how much of a perf impact these changes have. Would appreciate some more insight on whether these benchmarks are meaningful. 

cc @yuzefovich, your thoughts on these benchmarks would be great as well as these changes impact the query hot path (Would appreciate your thoughts on this PR generally as well if you have time. The PR looks big but most of the logic is contained within `sql/audit_logging.go` and the `auditlogging` package).