@knz 2023-05-08
I would recommend:

1. indeed, use a single node cluster.
2. perform the workload initialization just once; make a copy of the data files as baseline; then run the benchmark twice (with the same initialization data, restored from the file copy) with the two different versions of the code.