@THardy98 2023-05-17
Here are the benchmarking results using a single-node cluster, using the same workload initialization. This comparison was between the 23.1 release branch and a 23.1 backport of these changes:

```
thomashardy@gceworker-thomashardy:~/go/src/github.com/cockroachdb/cockroach$ ./bin/benchstat release23-1-benchstat-input.txt audit-logging-benchstat-input.txt
name   old s/op   new s/op    delta
[q1]   30.7 ±56%   30.4 ±59%     ~     (p=0.681 n=35+34)
[q2]   18.3 ±33%   18.3 ±37%     ~     (p=0.939 n=33+33)
[q3]   21.5 ±50%   20.0 ±30%     ~     (p=0.068 n=34+32)
[q4]   7.99 ±18%   7.61 ±27%     ~     (p=0.180 n=34+33)
[q5]   13.0 ±23%   14.0 ±27%   +7.29%  (p=0.014 n=33+33)
[q6]   31.0 ±38%   32.9 ±39%     ~     (p=0.130 n=33+32)
[q7]   33.8 ±24%   33.3 ±19%     ~     (p=0.846 n=32+32)
[q8]   4.86 ±20%   4.96 ±32%     ~     (p=0.472 n=32+32)
[q9]   43.6 ± 9%   41.1 ±22%   -5.79%  (p=0.001 n=32+31)
[q10]  10.6 ±18%    9.9 ±18%   -7.20%  (p=0.014 n=32+29)
[q11]  4.88 ±28%   4.61 ±26%     ~     (p=0.137 n=32+30)
[q12]  32.8 ±36%   35.7 ±34%   +8.86%  (p=0.022 n=31+30)
[q13]  8.80 ±18%   8.84 ±28%     ~     (p=0.953 n=31+29)
[q14]  3.50 ±31%   3.41 ±29%     ~     (p=0.360 n=31+30)
[q15]  60.0 ±95%  72.2 ±105%  +20.40%  (p=0.041 n=25+27)
[q16]  5.42 ±26%   5.26 ±46%     ~     (p=0.810 n=26+27)
[q17]  1.73 ±43%   1.76 ±32%     ~     (p=0.919 n=29+26)
[q18]  22.3 ±37%   23.2 ±32%     ~     (p=0.335 n=27+26)
[q19]  2.27 ±34%   2.46 ±61%     ~     (p=0.210 n=26+27)
[q20]  61.3 ± 9%   66.2 ±12%   +7.94%  (p=0.000 n=24+23)
[q21]  31.2 ±11%   30.3 ±18%     ~     (p=0.136 n=26+26)
[q22]  3.03 ±25%   3.17 ±34%     ~     (p=0.183 n=27+27)
```