@yuzefovich 2023-05-18
This comparison looks like it was of TPCH queries, ran one at a time with concurrency of 1 (right?). This is probably not that interesting for observing the impact of a change on the "hot path" since TPCH queries are analytical in nature, so they process a lot of data, which makes any overhead introduced by the change you want to benchmark become insignificant in comparison to the miscellaneous SQL operations (like joins and aggregation). I'd recommend using a very simple workload, perhaps KV95, for the benchmarking here - the idea is that we want to make the query processing as small as possible so that the overhead of this PR (if there is any) becomes more pronounced.