@yuzefovich 2023-05-20
> I'm not sure what to make of these results on my own (i.e. how much of a regression this is, is this bad?), nor how to make use of these outputs with benchstat.

Yeah, the impact is quite severe. The way to compare things is to pick one single stat (e.g. `ops/sec(cum)` from `result` line) and then compare between two outputs:
- before `7970.3`
- after `6416.3`

which roughly translates to 20% reduction in throughput. Granted, this is probably the largest perf hit we could find because KV is very simply workload, on more complicated ones the number will be smaller; still, 20% seems to me as a definitely too big to accept.

I don't have the necessary context on the feature to understand whether the customers that want this will accept it or whether such hit is expected. My intuition says that we'd want to limit this to 5% hit on KV95 workload. Consider taking CPU profiles while running KV95 before and after this change to see what stands out as the main cause for the regression.